Time: 88 ms (71.01%) | Memory: 27.2 MB (31.24%) - LeetSync

+++ 1615-range-sum-of-sorted-subarray-sums/range-sum-of-sorted-subarray-sums.cpp
@@ -12,7 +12,7 @@ public:
             }
         }
         sort(arr.begin(),arr.end());
-        long long int ans=0;
+        long int ans=0;
         for(int i=left-1;i<=right-1;i++){
             ans+=arr[i];
         }
